Refactor post-processing functions into own class
